+++ src/components/components.ts
@@ -1,4 +1,4 @@
-/* Web components. File generated at Thu, 30 Jul 2026 22:01:54 GMT. */
+/* Web components. File generated at Sat, 01 Aug 2026 18:46:14 GMT. */
 export * from './collapsible-content-panel/collapsible-content-panel.component';
 export * from './fixed-scroll-panel/fixed-scroll-panel.component';
 export * from './command-prompt/command-prompt.component';